Enhance documentation and refactor request and response handling for clarity and maintainability
