+++ ripress/src/tests/response/mod.rs
@@ -12,6 +12,35 @@ mod status_code;
 mod streaming_test;
 
 #[cfg(test)]
+mod test {
+    use crate::res::response_cookie::Cookie;
+
+    impl Cookie {
+        pub(crate) fn is_add_cookie(&self) -> bool {
+            matches!(self, Cookie::AddCookie(_))
+        }
+
+        pub(crate) fn is_remove_cookie(&self) -> bool {
+            matches!(self, Cookie::RemoveCookie(_))
+        }
+
+        pub fn value(&self) -> &str {
+            match self {
+                Cookie::AddCookie(add_cookie) => add_cookie.value,
+                Cookie::RemoveCookie(name) => name,
+            }
+        }
+
+        pub fn name(&self) -> &str {
+            match self {
+                Cookie::AddCookie(add_cookie) => add_cookie.name,
+                Cookie::RemoveCookie(name) => name,
+            }
+        }
+    }
+}
+
+#[cfg(test)]
 impl HttpResponse {
     pub(crate) fn get_status_code(&self) -> u16 {
         self.status_code.as_u16()
